+++ src/agentscope/codeact/__init__.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
 """This module provides CodeAct problem solving ability to agent."""
 
-from .code_act_run_tool_server import (
-    CodeActRunToolServer,
-    RunToolRequest,
+from .code_act_tool_call_server import (
+    CodeActToolCallServer,
+    ToolCallRequest,
 )
 
 
 __all__ = [
-    "CodeActRunToolServer",
-    "RunToolRequest",
+    "CodeActToolCallServer",
+    "ToolCallRequest",
 ]
